requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
filelock==3.19.1
flasgger==0.9.7.1
Flask==3.1.2
frozenlist==1.7.0
fsspec==2025.9.0
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.2
mistune==3.1.4
mpmath==1.3.0
multidict==6.6.4
mysql-connector-python==8.0.33
networkx==3.5
numpy==1.26.4
openai==0.28.0
openpyxl==3.1.5
packaging==25.0
pandas==2.1.4
pillow==11.3.0
propcache==0.3.2
protobuf==3.20.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.57.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
